server/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
ultralytics==8.3.40
opencv-python-headless==4.10.0.84
python-multipart==0.0.18
numpy==1.26.4
python-dotenv==1.0.1
